Create CreateDatabase.py
This file creates the database of scales that ScaleFinder.py searches for. Instead of hardcoding all the scales into the ScaleFinder.py file (which would add a lot of unnecessary code to the main file) I broke up the program into two separate files. This code only needs to be run once, and will create a file in your directory called Scales. 

Uses the module shelve